rel 4.0.0

+++ syslibs/cpm/fflush.c
@@ -27,15 +27,18 @@
 
 int fflush(register FILE * f)
 {
-	unsigned	cnt;
+	uint16_t cnt;
 
-	if(!(f->_flag & _IOWRT) || f->_base == (char *)NULL || (cnt = BUFSIZ - f->_cnt) == 0)
+	if (!(f->_flag & _IOWRT) || f->_base == (char *)NULL || (cnt = BUFSIZ - f->_cnt) == 0)
 		return 0;
-	if(write(fileno(f), f->_base, cnt) != cnt)
+
+	if (write(fileno(f), f->_base, cnt) != cnt)
 		f->_flag |= _IOERR;
+	
 	f->_cnt = BUFSIZ;
 	f->_ptr = f->_base;
-	if(f->_flag & _IOERR)
+	if (f->_flag & _IOERR)
 		return(EOF);
+	
 	return 0;
 }
